+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV010 - 03/04/2023 - Add the AESCommunicator class.
 - LASRV010 - 03/04/2023 - Add the RequestManager class.
 - LASRV010 - 03/04/2023 - Add the RequestCPU class.
+- LASRV010 - 03/04/2023 - Add the RequestRAM class.
 
 ### Changed
 - LASRV010 - 03/04/2023 - Changes in Request class.